Fixed reference to API keys and storage directory
Local storage must be updated when making it a package. I.e the path must be able to be set from the command line as a local environment variable or somehow otherwise it won't work for individual users as an import. Alternatives include managing a SQL database online whcih is takes from ?? Only small ammount of infomration is required and the storage method could be done effectively with link acting as a hash